refactor(front): move workflow run step logs into workflow-actions
Removes the workflow-run/observability folder and colocates each step log
detail component with its action:

- ai-agent detail + tool call row -> ai-agent-action/components
- code detail -> code-action/components
- http request detail -> http-request-action/components
- dispatcher, entries, email detail and styles -> workflow-actions/components
- formatters split one export per file -> workflow-actions/utils
