+++ packages/twenty-front/src/modules/workflow/workflow-steps/workflow-actions/utils/formatBytes.ts
@@ -1,11 +1,3 @@
-export const formatDuration = (ms: number): string => {
-  if (ms < 1000) {
-    return `${ms}ms`;
-  }
-
-  return `${(ms / 1000).toFixed(1)}s`;
-};
-
 export const formatBytes = (bytes: number): string => {
   if (bytes < 1024) {
     return `${bytes} B`;
