+++ packages/twenty-front/src/modules/workflow/workflow-steps/workflow-actions/utils/formatDuration.ts
@@ -1,0 +1,7 @@
+export const formatDuration = (ms: number): string => {
+  if (ms < 1000) {
+    return `${ms}ms`;
+  }
+
+  return `${(ms / 1000).toFixed(1)}s`;
+};
